aegis/pyproject.toml
[tool.poetry]
name = "aegis-dashboard"
version = "0.1.0"
description = "Aegis - Enterprise Trading Analytics Dashboard"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
dash = "^2.9.0"
dash-bootstrap-components = "^1.3.0"
pandas = "^1.5.0"
numpy = "^1.23.0"
plotly = "^5.11.0"
scipy = "^1.9.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
redis = "^4.4.0"
python-dotenv = "^0.21.0"
pandas-ta = "^0.3.14"
empyrical = "^0.5.5"
pyfolio = "^0.9.2"
scikit-learn = "^1.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
